+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # ConfigTUI Changelog
 
+### [1.1] - 2023-06-06
+- Added highlighter to show changes [update/insert/delete]
+- Code optimizations under the hood for better performance
+
 ### [1.0] - 2023-06-05
 - Added options to add / delete nodes
 - Added CLI options